fix the circular dep in the anim folder
